create homework-4

+++ src/controllers/notesController.js
@@ -2,8 +2,10 @@ import { Note } from '../models/note.js';
 import createHttpError from 'http-errors';
 
 export const getAllNotes = async (req, res) => {
+  const { _id: userId } = req.user;
   const { page = 1, perPage = 10, tag, search } = req.query;
-  const filter = {};
+
+  const filter = { userId };
   if (tag) {
     filter.tag = tag;
   }
@@ -14,11 +16,9 @@ export const getAllNotes = async (req, res) => {
     ];
   }
   const skip = (page - 1) * perPage;
-  // const [notes, totalNotes] = await Promise.all([
-  //   Note.find().skip(skip).limit(perPage),
-  //   Note.countDocuments(),
-  // ]); якщо хочемо виконувати запити паралельно, а не послідовно і вони не залежать один від одного, то можна так, але в даному випадку це не критично, бо обидва запити швидкі і не навантажують базу даних
+
   const notes = await Note.find(filter).skip(skip).limit(perPage);
+
   const totalNotes = await Note.countDocuments(filter);
   const totalPages = Math.ceil(totalNotes / perPage);
   res.status(200).json({
@@ -32,8 +32,9 @@ export const getAllNotes = async (req, res) => {
 
 export const getNoteById = async (req, res) => {
   const { noteId } = req.params;
-  //const note = await Note.findOne({ _id: noteId });
-  const note = await Note.findById(noteId); //можна так, якщо noteId є валідним ObjectId, інакше буде помилка CastError
+  const { _id: userId } = req.user;
+  const note = await Note.findOne({ _id: noteId, userId });
+  //const note = await Note.findById(noteId); //можна так, якщо noteId є валідним ObjectId, інакше буде помилка CastError
   if (!note) {
     throw createHttpError(404, `Note with id ${noteId} not found`);
     //const error = new Error(`Note with id ${noteId} not found`);
@@ -44,16 +45,22 @@ export const getNoteById = async (req, res) => {
 };
 
 export const createNote = async (req, res) => {
-  const newNote = await Note.create(req.body);
+  const { _id: userId } = req.user;
+  const newNote = await Note.create({ ...req.body, userId });
   res.status(201).json(newNote);
 };
 
 export const updateNote = async (req, res) => {
   const { noteId } = req.params;
-  const updatedNote = await Note.findOneAndUpdate({ _id: noteId }, req.body, {
-    returnDocument: 'after',
-    runValidators: true,
-  });
+  const { _id: userId } = req.user;
+  const updatedNote = await Note.findOneAndUpdate(
+    { _id: noteId, userId },
+    req.body,
+    {
+      returnDocument: 'after',
+      runValidators: true,
+    },
+  );
   if (!updatedNote) {
     throw createHttpError(404, 'Note not found');
   }
@@ -62,7 +69,8 @@ export const updateNote = async (req, res) => {
 
 export const deleteNote = async (req, res) => {
   const { noteId } = req.params;
-  const deletedNote = await Note.findOneAndDelete({ _id: noteId });
+  const { _id: userId } = req.user;
+  const deletedNote = await Note.findOneAndDelete({ _id: noteId, userId });
   if (!deletedNote) {
     throw createHttpError(404, 'Note not found');
   }
